requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.12.1
appdirs==1.4.4
astor==0.8.1
asttokens==3.0.1
async-generator==1.10
attrs==25.4.0
autogen-agentchat==0.7.5
autogen-core==0.7.5
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blessed==1.30.0
blinker==1.9.0
brotli==1.2.0
build==1.4.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.1.8
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
crewai==1.9.3
cryptography==46.0.5
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==5.2.1
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
editor==1.7.0
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.129.0
ffmpy==1.0.0
filelock==3.24.2
FLAML==2.5.0
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
git-python==1.0.3
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.72.0
gradio==6.5.1
gradio_client==2.0.3
greenlet==3.3.1
groovy==0.1.2
grpcio==1.78.0
h11==0.16.0
html2image==2.0.7
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
identify==2.6.16
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
inquirer==3.4.1
instructor==1.12.0
ipykernel==6.31.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jinxed==1.3.0
jiter==0.10.0
joblib==1.5.3
json5==0.10.0
json_repair==0.25.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
kubernetes==35.0.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.13
langchain-experimental==0.4.1
langchain-text-splitters==1.1.0
langgraph==1.0.8
langgraph-sdk==0.3.6
langsmith==0.7.4
litellm==1.40.8
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mcp==1.23.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.2
nodeenv==1.10.0
numpy==1.26.4
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.24.1
open-interpreter==0.3.4
openai==1.83.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.11.7
ormsgpack==1.12.2
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==3.0.0
parso==0.8.6
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
platformdirs==4.9.2
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==5.29.6
psutil==5.9.8
pure_eval==0.2.3
pyautogen==0.2.35
pybase64==1.4.3
pycparser==3.0
pydantic==2.11.10
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
pypdfium2==5.4.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.22
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
readchar==4.2.1
referencing==0.37.0
regex==2024.9.11
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.30.0
runs==1.3.0
safehttpx==0.1.7
selenium==4.20.0
semantic-version==2.10.0
Send2Trash==1.8.3
setuptools==69.5.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
sse-starlette==3.2.0
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
tenacity==8.5.0
termcolor==3.3.0
tiktoken==0.6.0
tokenizers==0.20.3
tokentrim==0.1.13
toml==0.10.2
tomli==2.0.2
tomli_w==1.1.0
tomlkit==0.13.3
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
trio==0.33.0
trio-typing==0.10.0
trio-websocket==0.12.2
typer==0.12.0
typer-cli==0.12.0
typer-slim==0.12.0
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
uv==0.9.30
uvicorn==0.41.0
virtualenv==20.37.0
watchfiles==1.1.1
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.5
wget==3.2
wheel==0.46.3
wsproto==1.3.2
xmod==1.9.0
xxhash==3.6.0
yarl==1.22.0
yt-dlp==2026.2.4
zipp==3.23.0
zstandard==0.25.0
